renamed script optimization Attrs.h header to prepare for factoring large Compile.h
